Implement authorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\AskQuestionRequest;
 use App\Question;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionController extends Controller
 {
@@ -36,7 +37,7 @@ class QuestionController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  AskQuestionRequest  $request
+     * @param AskQuestionRequest $request
      * @return \Illuminate\Http\Response
      */
     public function store(AskQuestionRequest $request)
@@ -50,7 +51,7 @@ class QuestionController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Question  $question
+     * @param \App\Question $question
      * @return \Illuminate\Http\Response
      */
     public function show(Question $question)
@@ -65,23 +66,31 @@ class QuestionController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Question  $question
+     * @param \App\Question $question
      * @return \Illuminate\Http\Response
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
+
         return view('questions.edit', compact('question'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  AskQuestionRequest $request
-     * @param  \App\Question  $question
+     * @param AskQuestionRequest $request
+     * @param \App\Question $question
      * @return \Illuminate\Http\Response
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->update($request->only(['title', 'body']));
 
         return redirect('/questions')->with('success', 'Your question has been updated');
@@ -90,13 +99,17 @@ class QuestionController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Question  $question
+     * @param \App\Question $question
      * @return \Illuminate\Http\Response
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied");
+        }
+
         $question->delete();
 
-        return redirect('/questions')->with('success','Question has been deleted');
+        return redirect('/questions')->with('success', 'Question has been deleted');
     }
 }
